requirements.txt
flask
flask-session
anthropic
python-dotenv
python-docx
